+++ CHANGES.md
@@ -40,6 +40,8 @@ v0.10.0 (unreleased)
   and can be used to define logical relations between different attributes
   in a GUI-independent way.
 
+- Fix selections when using Matplotlib 2.x, PyQt5 and a retina display. [#1236]
+
 - Updated ComponentIDComboHelper to no longer store ``(cid, data)`` as the
   ``userData`` but instead just the ``cid`` (the data is now accessible via
   ``cid.parent``). [#1212]